+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.1.0-nullsafety.0
+
+- Add null safety support
+
+**Dev notes**: Switched to `mocktail` instead of `mockito` for tests because of
+null safety issues with mockito.
+
 ## 0.0.2
 
 - Fix issue with merge data json encoding